Move & split `FeedbackDataset`, and `pydantic.BaseModel`s
