+++ src/argilla/client/feedback/dataset.py
@@ -15,7 +15,6 @@
 import logging
 import tempfile
 from typing import TYPE_CHECKING, Any, Dict, Iterator, List, Optional, Tuple, Union
-from uuid import UUID
 
 try:
     from typing import Literal
@@ -24,17 +23,23 @@ except ImportError:
 
 from pydantic import (
     BaseModel,
-    Extra,
-    Field,
-    StrictInt,
-    StrictStr,
     ValidationError,
     create_model,
-    validator,
 )
 from tqdm import tqdm
 
 import argilla as rg
+from argilla.client.feedback.schemas import (
+    FIELD_TYPE_TO_PYTHON_TYPE,
+    FeedbackDatasetConfig,
+    FeedbackRecord,
+    FieldSchema,
+    OfflineFeedbackRecord,
+    QuestionSchema,
+    RatingQuestion,
+    TextField,
+    TextQuestion,
+)
 from argilla.client.sdk.v1.datasets import api as datasets_api_v1
 from argilla.utils.dependency import requires_version
 
@@ -53,114 +58,6 @@ FETCHING_BATCH_SIZE = 250
 PUSHING_BATCH_SIZE = 32
 
 _LOGGER = logging.getLogger(__name__)
-
-
-class ValueSchema(BaseModel):
-    value: Union[StrictStr, StrictInt]
-
-
-class ResponseSchema(BaseModel):
-    user_id: Optional[UUID] = None
-    values: Dict[str, ValueSchema]
-    status: Literal["submitted", "discarded"] = "submitted"
-
-    @validator("user_id", always=True)
-    def user_id_must_have_value(cls, v):
-        if not v:
-            _LOGGER.warning(
-                "`user_id` not provided, so it will be set to `None`. Which is not an"
-                " issue, unless if you're planning to log the response in Argilla, as "
-                " it will be automatically set to the active `user_id`."
-            )
-        return v
-
-
-class FeedbackRecord(BaseModel):
-    fields: Dict[str, str]
-    responses: Optional[Union[ResponseSchema, List[ResponseSchema]]] = None
-    external_id: Optional[str] = None
-
-    @validator("responses", always=True)
-    def responses_must_be_a_list(cls, v):
-        if not v:
-            return []
-        if isinstance(v, ResponseSchema):
-            return [v]
-        return v
-
-    class Config:
-        extra = Extra.forbid
-
-
-class OfflineFeedbackRecord(BaseModel):
-    id: Optional[str] = None
-    fields: Dict[str, str]
-    responses: List[ResponseSchema] = []
-    external_id: Optional[str] = None
-    inserted_at: Optional[str] = None
-    updated_at: Optional[str] = None
-
-
-class FieldSchema(BaseModel):
-    name: str
-    title: Optional[str] = None
-    required: Optional[bool] = True
-    settings: Dict[str, Any]
-
-    @validator("title", always=True)
-    def title_must_have_value(cls, v, values):
-        if not v:
-            return values["name"].capitalize()
-        return v
-
-
-class TextField(FieldSchema):
-    settings: Dict[str, Any] = Field({"type": "text"}, const=True)
-
-
-FIELD_TYPE_TO_PYTHON_TYPE = {"text": str}
-
-
-class QuestionSchema(BaseModel):
-    name: str
-    title: Optional[str] = None
-    description: Optional[str] = None
-    required: Optional[bool] = True
-    settings: Dict[str, Any]
-
-    @validator("title", always=True)
-    def title_must_have_value(cls, v, values):
-        if not v:
-            return values["name"].capitalize()
-        return v
-
-
-# TODO(alvarobartt): add `TextResponse` and `RatingResponse` classes
-class TextQuestion(QuestionSchema):
-    settings: Dict[str, Any] = Field({"type": "text"}, const=True)
-
-
-class RatingQuestion(QuestionSchema):
-    settings: Dict[str, Any] = Field({"type": "rating"})
-    values: List[int] = Field(unique_items=True)
-
-    @validator("values", always=True)
-    def update_settings_with_values(cls, v, values):
-        if v:
-            values["settings"]["options"] = [{"value": value} for value in v]
-        return v
-
-    class Config:
-        validate_assignment = True
-
-
-AllowedQuestionTypes = Union[TextQuestion, RatingQuestion]
-
-
-class FeedbackDatasetConfig(BaseModel):
-    fields: List[FieldSchema]
-    questions: List[AllowedQuestionTypes]
-    guidelines: Optional[str] = None
 
 
 class FeedbackDataset:
@@ -241,7 +138,7 @@ class FeedbackDataset:
         self,
         *,
         fields: List[FieldSchema],
-        questions: List[AllowedQuestionTypes],
+        questions: List[Union[TextQuestion, RatingQuestion]],
         guidelines: Optional[str] = None,
     ) -> None:
         """Initializes a `FeedbackDataset` instance locally.
